api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
anthropic>=0.49.0
python-dotenv>=1.0.0
pydantic>=2.0.0
